refactor: extract chat automation logic into useChatAutomation hook and streamline ChatState
