+++ src/chats/chat-state.tsx
@@ -2,15 +2,14 @@ import { aiFetchStreamingResponse } from '@/ai/fetch'
 import ChatUI from '@/components/chat/chat-ui'
 import { useThrottledCallback } from '@/hooks/use-throttle'
 import { trackEvent } from '@/lib/posthog'
-import { getTriggerPromptForThread } from '@/dal'
 import { useMCP } from '@/lib/mcp-provider'
 import type { SaveMessagesFunction, ThunderboltUIMessage } from '@/types'
 import { useChat, type UseChatHelpers } from '@ai-sdk/react'
-import { useQuery } from '@tanstack/react-query'
 import { DefaultChatTransport } from 'ai'
-import { useCallback, useEffect, useRef } from 'react'
+import { useCallback, useEffect } from 'react'
 import { v7 as uuidv7 } from 'uuid'
 import { useChatModel } from './use-chat-model'
+import { useChatAutomation } from './use-chat-automation'
 
 interface ChatStateProps {
   id: string
@@ -85,7 +84,7 @@ export default function ChatState({ id, initialMessages, saveMessages }: ChatSta
       trackEvent('chat_receive_reply', {
         model: selectedModelIdRef.current,
         length: message.parts?.reduce((acc, part) => acc + (part.type === 'text' ? part.text.length : 0), 0) || 0,
-        reply_number: chatMessages.length + 1,
+        reply_number: chatHelpers.messages.length + 1,
       })
     },
     onError: (error) => {
@@ -96,34 +95,7 @@ export default function ChatState({ id, initialMessages, saveMessages }: ChatSta
 
   useSavePartialAssistantMessages({ chatHelpers, id, saveMessages })
 
-  const { messages: chatMessages, status } = chatHelpers
-
-  // Load the automation prompt that triggered this chat, if any
-  const { data: triggerData } = useQuery({
-    queryKey: ['triggerPrompt', id],
-    queryFn: () => getTriggerPromptForThread(id),
-  })
-
-  // Auto-run assistant if thread ends with user message (e.g., automation) and no assistant response yet
-  const hasTriggeredRef = useRef(false)
-  useEffect(() => {
-    if (hasTriggeredRef.current) return
-
-    if (
-      selectedModelId &&
-      status === 'ready' &&
-      chatMessages.length > 0 &&
-      chatMessages[chatMessages.length - 1].role === 'user'
-    ) {
-      hasTriggeredRef.current = true
-      // Regenerate assistant response for the last user message
-      chatHelpers.regenerate().catch((err) => {
-        hasTriggeredRef.current = false
-        console.error('Auto regenerate error', err)
-      })
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [status, selectedModelId])
+  const { triggerData } = useChatAutomation({ chatHelpers, chatThreadId: id, selectedModelId })
 
   // If we don't pass a selectedModelId to the ChatUI, it will warn about changing an input from uncontrolled to controlled
   if (!selectedModelId) {
